refactor(playback::rusty_backend): move icy-metadata handling to own module
